ユーザーページ、日記詳細ページのハンドラ修正・不要な記述を削除
