+++ src/pages/user/[userId]/index.tsx
@@ -1,6 +1,5 @@
 import { NextPage } from "next";
 import { MyNextContext } from "next-redux-wrapper";
-import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { END } from "redux-saga";
@@ -13,20 +12,11 @@ import DiaryCard from "../../../components/DiaryCard";
 import EditButton from "../../../components/EditButton";
 import Heading from "../../../components/Heading";
 import Layout from "../../../components/Layout";
-import Modal from "../../../components/Modal";
-import PageBottomNotifier, {
-  NotifierStatus,
-} from "../../../components/PageBottomNotifier";
 import UserProfile from "../../../components/UserProfile";
 import BreakPoint from "../../../constants/BreakPoint";
 import { getUserFromFirestore } from "../../../lib/firestore";
 import { RootState, wrapper } from "../../../store";
-import {
-  createDraft,
-  deleteDiary,
-  getDiaries,
-  setDiaries,
-} from "../../../store/diary/actions";
+import { getDiaries, setDiaries } from "../../../store/diary/actions";
 import { Diary } from "../../../store/diary/types";
 import { userSignIn } from "../../../store/user/actions";
 import { User } from "../../../store/user/types";
@@ -72,7 +62,6 @@ type UserPageProps = {
 };
 
 const UserPage: NextPage<UserPageProps> = ({ author, user }: UserPageProps) => {
-  const router = useRouter();
   const dispatch = useDispatch();
 
   const diaries = useSelector<RootState, Diary[]>(
@@ -83,10 +72,6 @@ const UserPage: NextPage<UserPageProps> = ({ author, user }: UserPageProps) => {
   const [unsubscribeDb, setUnsubscribeDb] = useState<{
     [key: string]: (() => void) | undefined;
   }>({});
-
-  const [modalId, setModalId] = useState("");
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [notifierStatus, setNotifierStatus] = useState("hidden");
 
   const addDbListener = (id: string) => {
     const listener = firestore.collection(`users/${id}/diaries`).onSnapshot(
@@ -114,16 +99,6 @@ const UserPage: NextPage<UserPageProps> = ({ author, user }: UserPageProps) => {
     if (unsubscribeDb.listener) {
       unsubscribeDb.listener();
     }
-  };
-
-  const handleOpenDeleteModal = (id: string) => {
-    setModalId(id);
-    setIsModalOpen(true);
-  };
-
-  const handleAfterModalClose = async () => {
-    setNotifierStatus("visible" as NotifierStatus);
-    setTimeout(() => setNotifierStatus("hidden" as NotifierStatus), 1000);
   };
 
   useEffect(() => {
@@ -162,21 +137,6 @@ const UserPage: NextPage<UserPageProps> = ({ author, user }: UserPageProps) => {
             <NoDiaryText>まだ日記はありません</NoDiaryText>
           )}
           <StyledEditButton />
-          <Modal.ConfirmDelete
-            id={modalId}
-            isOpen={isModalOpen}
-            onRequestClose={() => setIsModalOpen(false)}
-            onAfterClose={handleAfterModalClose}
-            onDelete={() => {
-              dispatch(
-                deleteDiary({ firestore, userId: author.uid, diaryId: modalId })
-              );
-            }}
-          />
-          <PageBottomNotifier
-            text="日記を削除しました"
-            status={notifierStatus as NotifierStatus}
-          />
         </>
       )}
       {user?.uid === author.uid && (
